swetha playto/backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
